SPY-197: Add a flag for json compression
Motivation
----------
We would like the option to enable compression on json objects without
having to rewrite the default transcoder

Modifications
-------------

Added a boolean flag, skipJsonCompression, that is true by default
along with a corresponding test for when it is false. I did not
add the modifications to the other transcoders but can if the is
prefered. Additionally I am not familiar with the test suite and
not sure if the test needs modification.

Result
------

Default behavior remains unchanged and the behavior before SPY-58 can
be achieved through configuration.

Change-Id: I34402745671c8c751b4d76955beb6304280bea42